requirements.txt
filetype==1.2.0
pytesseract==0.3.10
Pillow~=9.5.0
